Refactor simple hydrides

+++ crates/blue_book/src/chapters/p_1_general/p_14_general_rules/p_14_3_locants.rs
@@ -5,6 +5,6 @@ use crate::Element;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Locant {
     Unspecified,
-    Number(u8),
-    Element(u8, Element),
+    Number(u16),
+    Element(u16, Element),
 }
